+++ src/nodejs/lambda-handlers/questions.js
@@ -17,13 +17,18 @@ async function hasPerms(uid, perms) {
   return false;
 }
 
-async function deleteInput(params){
-  const question = await db.question.findById({id: params.questionId})
-  const curInput = question.inputs.map(input => input.control_id)
-  const toDelete = curInput.filter(cid => !params.inputs.some(testInput => testInput.control_id === cid))
-  console.log(params.inputs.length)
-  console.log(toDelete)
-  return await db.question.deleteInput({toDelete})
+async function deleteInput(params) {
+  const question = await db.question.findById({ id: params.questionId });
+
+  const toDelete = [];
+  for (let i = 0; i < params.inputs.length; i += 1) {
+    params.inputs[i].list_order = i;
+    question.inputs.forEach((quesInput) => {
+      // eslint-disable-next-line
+      quesInput.control_id !== params.inputs[i].control_id ? toDelete.push(quesInput.control_id) : '';
+    });
+  }
+  return db.question.deleteInput({ toDelete });
 }
 
 async function findAllMethod({ params, context }) {
@@ -48,15 +53,15 @@ async function addMethod({ params, context }) {
 
 async function updateInputsMethod({ params, context }) {
   if (await hasPerms(context.user_id, editPerms)) {
-    const deletes = await deleteInput(params)
-    console.log('did enter')
+    await deleteInput(params);
     const promises = params.inputs.map(async (inputElem) => db.question.updateInput(
       {
         input: inputElem,
         questionId: params.questionId
       }
     ));
-    const updates =  Promise.all(promises);
+    Promise.all(promises);
+    return (db.question.findById(params));
   }
   return {};
 }